Move promotion checking into its own controller
