+++ app/views/fragments/checkout/promoCode.scala.html
@@ -1,7 +1,7 @@
 @import com.gu.i18n.Country
 @import com.gu.memsub.Subscription.ProductRatePlanId
 @import com.gu.memsub.promo.PromoCode
-@import controllers.routes.Checkout.validatePromoCode
+@import controllers.routes.Promotion.validatePromoCode
 @(promoCode: Option[PromoCode])
 @if(false) {
     <div class="promo-code-applied">✓&nbsp; Promotion applied</div>
