+++ conf/routes
@@ -22,7 +22,7 @@ POST        /checkout                        controllers.Checkout.handleCheckout
 GET         /checkout/check-identity         controllers.Checkout.checkIdentity(email: String)
 POST        /checkout/check-account          controllers.Checkout.checkAccount
 POST        /checkout/register-guest-user    controllers.Checkout.convertGuestUser
-GET         /checkout/lookupPromotion        controllers.Checkout.validatePromoCode(promoCode: PromoCode, productRatePlanId: ProductRatePlanId, country: Country)
+GET         /checkout/lookupPromotion        controllers.Promotion.validatePromoCode(promoCode: PromoCode, productRatePlanId: ProductRatePlanId, country: Country)
 GET         /checkout/validateDelivery       controllers.Checkout.validateDelivery(postCode: String)
 GET         /checkout/findAddress            controllers.Checkout.findAddress(postCode: String)
